# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'ToggleBox: Remote Config & Feature Flags API'
  description: |
    **ToggleBox** is a multi-platform remote configuration and feature flag management service.

    ## Key Features

    - **Remote Configuration**: Version-controlled, immutable config management
    - **Feature Flags**: Flexible toggles with percentage and targeted rollouts
    - **Multi-Platform**: Separate configs for web, mobile, and other platforms
    - **Environment Management**: Staging, production, and custom environments
    - **CDN Caching**: CloudFront-backed caching with cache invalidation
    - **Client SDKs**: JavaScript, Next.js, and Expo/React Native

    ## API Structure

    ### Public Endpoints (Read-Only)
    - **Path**: `/api/v1/platforms/*`
    - **Access**: Publicly accessible, no authentication
    - **Methods**: GET only
    - **Caching**: CloudFront CDN with configurable TTL

    ### Authentication Endpoints
    - **Paths**: `/api/v1/auth/*`, `/api/v1/users/*`, `/api/v1/api-keys/*`
    - **Access**: Public for registration/login, authenticated for management
    - **Methods**: POST, GET, PATCH, DELETE

    ### Internal Endpoints (Write Operations)
    - **Path**: `/api/v1/internal/*`
    - **Access**: Network-restricted or authenticated
    - **Methods**: POST, PUT, PATCH, DELETE
    - **Security**: JWT Bearer tokens or API Keys required

    ## Authentication

    **JWT Bearer Tokens** (user accounts):
    - Register: POST `/api/v1/auth/register`
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /internal/platforms/{platform}/environments/{environment}/configs/{configKey}/stats:
    get:
      tags:
        - Stats
      summary: Get config statistics
      description: Get usage statistics for a configuration
      operationId: getConfigStats
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - name: configKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Config statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ConfigStats'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/flags/{flagKey}/stats:
    get:
      tags:
        - Stats
      summary: Get flag statistics
      description: Get usage statistics for a feature flag
      operationId: getFlagStats
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        '200':
          description: Flag statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FlagStats'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/flags/{flagKey}/stats/by-country:
    get:
      tags:
        - Stats
      summary: Get flag statistics by country
      description: Get usage statistics for a feature flag grouped by country
      operationId: getFlagStatsByCountry
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        '200':
          description: Flag statistics by country
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CountryStats'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/flags/{flagKey}/stats/daily:
    get:
      tags:
        - Stats
      summary: Get flag statistics daily time series
      description: Get daily usage statistics for a feature flag
      operationId: getFlagStatsDaily
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Number of days to retrieve
      responses:
        '200':
          description: Daily flag statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DailyStats'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/experiments/{experimentKey}/stats:
    get:
      tags:
        - Stats
      summary: Get experiment statistics
      description: Get usage statistics for an experiment
      operationId: getExperimentStats
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/ExperimentKey'
      responses:
        '200':
          description: Experiment statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ExperimentStats'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/experiments/{experimentKey}/conversions:
    post:
      tags:
        - Stats
      summary: Record experiment conversion
      description: Record a conversion event for an experiment
      operationId: recordConversion
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/ExperimentKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - metricKey
              properties:
                userId:
                  type: string
                metricKey:
                  type: string
                value:
                  type: number
                  description: Numeric value for revenue/count metrics
      responses:
        '200':
          description: Conversion recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # CACHE INVALIDATION
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication
  parameters:
    Platform:
      name: platform
      in: path
      required: true
      schema:
        type: string
      description: Platform identifier
...
    Environment:
      name: environment
      in: path
      required: true
      schema:
        type: string
      description: Environment name
...
    FlagKey:
      name: flagKey
      in: path
      required: true
      schema:
        type: string
      description: Feature flag key identifier
    ExperimentKey:
      name: experimentKey
      in: path
      required: true
      schema:
        type: string
      description: Experiment key identifier
  schemas:
...
    ConfigStats:
      type: object
      properties:
        configKey:
          type: string
        totalFetches:
          type: integer
        uniqueUsers:
          type: integer
        lastFetchedAt:
          type: string
          format: date-time
    FlagStats:
      type: object
      properties:
        flagKey:
          type: string
        totalEvaluations:
          type: integer
        uniqueUsers:
          type: integer
        valueACount:
          type: integer
        valueBCount:
          type: integer
        lastEvaluatedAt:
          type: string
          format: date-time
    CountryStats:
      type: object
      properties:
        country:
          type: string
        totalEvaluations:
          type: integer
        valueACount:
          type: integer
        valueBCount:
          type: integer
    DailyStats:
      type: object
      properties:
        date:
          type: string
          format: date
        totalEvaluations:
          type: integer
        uniqueUsers:
          type: integer
        valueACount:
          type: integer
        valueBCount:
          type: integer
    ExperimentStats:
      type: object
      properties:
        experimentKey:
          type: string
        totalAssignments:
          type: integer
        uniqueUsers:
          type: integer
        variationStats:
          type: array
          items:
            type: object
            properties:
              variationKey:
                type: string
              assignments:
                type: integer
              conversions:
                type: integer
              conversionRate:
                type: number
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
        error:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...